Fix linters errors.
