+++ Main.rb
@@ -1,16 +1,13 @@
-require './MyList'
+require './my_list'
 list = MyList.new(1, 2, 3, 4)
 
 # Test #all?
-list.all? {|e| e < 5}
-list.all? {|e| e > 5}
-
+puts(list.all? { |e| e < 5 })
+puts(list.all? { |e| e > 5 })
 
 # Test #any?
-list.any? {|e| e == 2}
-
-list.any? {|e| e == 5}
-
+puts(list.any? { |e| e == 2 })
+puts(list.any? { |e| e == 5 })
 
 # Test #filter
-list.filter {|e| e.even?}
+puts(list.filter(&:even?))
