--- MyList.rb
@@ -1,14 +1,0 @@
-require './MyEnumerable'
-
-class MyList
-    include MyEnumerable
-
-    def initialize(list)
-        @list = list
-    end
-    
-    def each(&block)
-        @list.each(&block)
-    end
-     
-end
